requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
mdutils==1.6.0
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
slack_sdk==3.27.1
smmap==5.0.1
tzdata==2024.1
urllib3==2.2.1
